+++ spec/javascripts/controller.spec.js
@@ -53,4 +53,38 @@ describe("marionette controller", function(){
     });
   });
 
+  describe("when closing a controller", function(){
+    var controller, closeHandler;
+
+    beforeEach(function(){
+      controller = new (Marionette.Controller.extend({
+        onClose: jasmine.createSpy("onClose")
+      }));
+
+      closeHandler = jasmine.createSpy("close");
+      controller.on("close", closeHandler);
+
+      spyOn(controller, "unbindAll").andCallThrough();
+      spyOn(controller, "unbind").andCallThrough();
+
+      controller.close();
+    });
+
+    it("should unbindAll events", function(){
+      expect(controller.unbindAll).toHaveBeenCalled();
+    });
+
+    it("should unbind events", function(){
+      expect(controller.unbind).toHaveBeenCalled();
+    });
+
+    it("should trigger a close event", function(){
+      expect(closeHandler).toHaveBeenCalled();
+    });
+
+    it("should call an onClose method", function(){
+      expect(controller.onClose).toHaveBeenCalled();
+    });
+  });
+
 });