+++ src/marionette.controller.js
@@ -5,6 +5,7 @@
 // modules and routers, and as a mediator for workflow
 // and coordination of other objects, views, and more.
 Marionette.Controller = function(options){
+  this.triggerMethod = Marionette.triggerMethod;
   this.options = options || {};
 
   Marionette.addEventBinder(this);
@@ -20,4 +21,10 @@ Marionette.Controller.extend = Marionette.extend;
 // --------------
 
 // Ensure it can trigger events with Backbone.Events
-_.extend(Marionette.Controller.prototype, Backbone.Events);
+_.extend(Marionette.Controller.prototype, Backbone.Events, {
+  close: function(){
+    this.unbindAll();
+    this.triggerMethod("close");
+    this.unbind();
+  }
+});